refactor #156 : 전역 상태에 따라 로딩 UI 또는 업로드 UI 분기 처리
