+++ src/features/upload/components/upload/upload.styles.ts
